refactor(storywrapper): rename scss files to use .module.scss
